+++ src/bldprogs/VBoxEditCoffLib.cpp
@@ -4,15 +4,25 @@
  */
 
 /*
- * Copyright (C) 2006-2020 Oracle Corporation
+ * Copyright (C) 2006-2022 Oracle and/or its affiliates.
  *
- * This file is part of VirtualBox Open Source Edition (OSE), as
- * available from http://www.virtualbox.org. This file is free software;
- * you can redistribute it and/or modify it under the terms of the GNU
- * General Public License (GPL) as published by the Free Software
- * Foundation, in version 2 as it comes in the "COPYING" file of the
- * VirtualBox OSE distribution. VirtualBox OSE is distributed in the
- * hope that it will be useful, but WITHOUT ANY WARRANTY of any kind.
+ * This file is part of VirtualBox base platform packages, as
+ * available from https://www.virtualbox.org.
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU General Public License
+ * as published by the Free Software Foundation, in version 3 of the
+ * License.
+ *
+ * This program is distributed in the hope that it will be useful, but
+ * WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, see <https://www.gnu.org/licenses>.
+ *
+ * SPDX-License-Identifier: GPL-3.0-only
  */
 
 
@@ -240,7 +250,8 @@ static int SelectMember(const char *pszEndsWith)
         }
 
         if (g_cVerbosity > 2)
-            fprintf(stderr, "debug: %#08x: %#010x %*.*s\n", off, cbFile - cbExtra, cchName, cchName, pchName);
+            fprintf(stderr, "debug: %#08x: %#010x %*.*s\n",
+                    (unsigned)off, (unsigned)(cbFile - cbExtra), (int)cchName, (int)cchName, pchName);
 
         /*
          * Do matching.
@@ -252,7 +263,7 @@ static int SelectMember(const char *pszEndsWith)
             g_cbMember = (unsigned)(cbFile - cbExtra);
             if (g_cVerbosity > 1)
                 fprintf(stderr, "debug: selected '%*.*s': %#x LB %#x\n",
-                        cchName, cchName, pchName, off + sizeof(*pHdr) + cbExtra, g_cbMember);
+                        (int)cchName, (int)cchName, pchName, (unsigned)(off + sizeof(*pHdr) + cbExtra), g_cbMember);
             return 0;
         }
 
@@ -312,7 +323,7 @@ static int RedefineSymbol(const char *pszOldEqualNew)
 
     if (g_cVerbosity > 2)
         fprintf(stderr, "debug: redefining symbol '%*.*s' to '%*.*s'...\n",
-                cchOld, cchOld, pszOld, cchNew, cchNew, pszNew);
+                (int)cchOld, (int)cchOld, pszOld, (int)cchNew, (int)cchNew, pszNew);
 
     /*
      * Parse COFF header.
